+++ fastapi_restly/db/_globals.py
@@ -18,6 +18,7 @@ class RestlyContext:
         "make_session",
         "session_generator",
         "sync_session_generator",
+        "warn_on_misuse",
         "warn_on_uncommitted",
     )
 
@@ -27,6 +28,7 @@ class RestlyContext:
     make_session: sessionmaker[Any] | None
     session_generator: Callable[[], AsyncIterator[SA_AsyncSession]] | None
     sync_session_generator: Callable[[], Iterator[SA_Session]] | None
+    warn_on_misuse: bool
     warn_on_uncommitted: bool
 
     def __init__(self) -> None:
@@ -36,6 +38,7 @@ class RestlyContext:
         self.make_session = None
         self.session_generator = None
         self.sync_session_generator = None
+        self.warn_on_misuse = False
         self.warn_on_uncommitted = True
 
     def __enter__(self) -> "RestlyContext":